+++ tests/reconfiguration.py
@@ -259,6 +259,8 @@ def test_add_node_endorsements_endpoints(network, args):
         (["AMD:kdsintf.amd.com"], True),
         (["AMD:invalid.amd.com"], False),
         (["Azure:invalid.azure.com", "AMD:kdsintf.amd.com"], True),  # Fallback server
+        # Won't work yet, see #5852
+        # (["THIM:$Fabric_NodeIPOrFQDN:2377"], True),
     ]
 
     for servers, expected_result in test_vectors: